프로젝트(Project) 선택 화면 작성 - 프로젝트 선택시 마지막 파일을 자동 선택하게끔 구성   - 이후 기획에 따라 수정될 예정
설정(Config)파일을 위한 Serialization 라이브러리 도입
사전설정(Workflow)파일 기획 필요

+++ composeApp/src/jvmMain/kotlin/com/ninetag/machum/external/FileManager.jvm.kt
@@ -14,11 +14,12 @@ internal actual suspend fun FileManager.createFolder(
 ): PlatformFile? = withContext(Dispatchers.IO) {
     try {
         val directory = PlatformFile(parentDirectory, name)
+        if (directory.exists()) return@withContext directory
         directory.createDirectories()
         directory
     } catch (e: Exception) {
         println("폴더 생성 실패: $e")
-        null
+        throw e
     }
 }
 
@@ -29,6 +30,7 @@ internal actual suspend fun FileManager.createFile(
     try {
         if (!parentDirectory.exists()) return@withContext null
         val newFile = parentDirectory / "$name.md"
+        if (newFile.exists()) return@withContext newFile
         newFile.writeString("")
         newFile
     } catch (e: Exception) {
@@ -36,3 +38,22 @@ internal actual suspend fun FileManager.createFile(
         null
     }
 }
+
+internal actual suspend fun FileManager.createConfig(
+    parentDirectory: PlatformFile
+): PlatformFile? = withContext(Dispatchers.IO) {
+    try {
+        if (!parentDirectory.exists()) return@withContext null
+        val newFile = parentDirectory / ".machum.json"
+        if (newFile.exists()) return@withContext newFile
+        newFile.writeString("")
+        newFile
+    } catch (e: Exception) {
+        println("Config 생성 실패: $e")
+        throw e
+    }
+}
+
+internal actual suspend fun FileManager.validPermission(file: PlatformFile): Boolean = true
+
+internal actual fun String.forPlatformFile(): String = this
